feat: Refactor payment models and repository to use new namespace; implement PaymentRequest and PaymentResponse contracts
