+++ services/payments/Payment.Api/Contracts/Requests/PaymentRequest.cs
@@ -1,0 +1,7 @@
+namespace Payment.Api.Contracts.Requests;
+public record PaymentRequest(
+    Guid OrderId,
+    Guid CustomerId,
+    decimal Price,
+    string Currency,
+    string PaymentMethod);
