+++ services/payments/Payment.Api/Contracts/Responses/PaymentResponse.cs
@@ -1,0 +1,7 @@
+namespace Payment.Api.Contracts.Responses;
+
+public record PaymentResponse(
+    int Id,
+    decimal Amount,
+    string Currency,
+    string PaymentMethod);
