+++ services/payments/Payment.Api/Data/Repositories/IPaymentRepository.cs
@@ -1,12 +1,11 @@
-namespace Payment.Api.Data.Repositories
+namespace Payment.Api.Data.Repositories;
+
+public interface IPaymentRepository
 {
-    public interface IPaymentRepository
-    {
-        // Define your repository methods here
-        Task<Models.Entities.Payment> GetPaymentByIdAsync(Guid id);
-        Task<IEnumerable<Models.Entities.Payment>> GetAllPaymentsAsync();
-        Task AddPaymentAsync(Models.Entities.Payment payment);
-        Task UpdatePaymentAsync(Models.Entities.Payment payment);
-        Task DeletePaymentAsync(Guid id);
-    }
+    // Define your repository methods here
+    Task<Entities.Payment> GetPaymentByIdAsync(Guid id);
+    Task<IEnumerable<Entities.Payment>> GetAllPaymentsAsync();
+    Task AddPaymentAsync(Entities.Payment payment);
+    Task UpdatePaymentAsync(Entities.Payment payment);
+    Task DeletePaymentAsync(Guid id);
 }
