--- services/payments/Payment.Api/Models/Entities/Payment.cs
@@ -1,13 +1,0 @@
-namespace Payment.Api.Models.Entities
-{
-    public class Payment
-    {
-        public Guid Id { get; set; }
-        public Guid OrderId { get; set; }
-        public Guid CustomerId { get; set; }
-        public decimal Price { get; set; }
-        public string Currency { get; set; } = null!;
-        public string PaymentMethod { get; set; } = null!;
-        public DateTime CreatedAt { get; set; }
-    }
-}
