--- services/payments/Payment.Api/Models/Requests/PaymentRequest.cs
@@ -1,9 +1,0 @@
-namespace Payment.Api.Models.Requests
-{
-    public record PaymentRequest(
-        Guid OrderId,
-        Guid CustomerId,
-        decimal Price,
-        string Currency,
-        string PaymentMethod);
-}
