--- services/payments/Payment.Api/Models/Responses/PaymentResponse.cs
@@ -1,4 +1,0 @@
-namespace Payment.Api.Models.Responses
-{
-    public record PaymentResponse(int Id, decimal Amount, string Currency, string PaymentMethod);
-}
