+++ services/payments/Payment.Api/Services/Interfaces/IPaymentService.cs
@@ -2,10 +2,10 @@ namespace Payment.Api.Data.Services.Interfaces
 {
     public interface IPaymentService
     {
-        Task<Models.Entities.Payment> GetPaymentByIdAsync(Guid id);
-        Task<IEnumerable<Models.Entities.Payment>> GetAllPaymentsAsync();
-        Task AddPaymentAsync(Models.Entities.Payment payment);
-        Task UpdatePaymentAsync(Models.Entities.Payment payment);
+        Task<Entities.Payment> GetPaymentByIdAsync(Guid id);
+        Task<IEnumerable<Entities.Payment>> GetAllPaymentsAsync();
+        Task AddPaymentAsync(Entities.Payment payment);
+        Task UpdatePaymentAsync(Entities.Payment payment);
         Task DeletePaymentAsync(Guid id);
     }
 }
